v1.0.10 Added files distribution
