+++ lib/templates/middleAssignment/star_rating.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 
 typedef RatingChangeCallback = void Function(double rating);
@@ -10,7 +8,12 @@ class StarRating extends StatelessWidget {
   final RatingChangeCallback onRatingChanged;
   final Color color;
 
-  const StarRating({super.key, this.starCount = 5, this.rating = .0, required this.onRatingChanged, required this.color});
+  const StarRating(
+      {super.key,
+      this.starCount = 5,
+      this.rating = .0,
+      required this.onRatingChanged,
+      required this.color});
 
   Widget buildStar(BuildContext context, int index) {
     Icon icon;
@@ -19,8 +22,7 @@ class StarRating extends StatelessWidget {
         Icons.star_border,
         color: Colors.yellow,
       );
-    }
-    else if (index > rating - 1 && index < rating) {
+    } else if (index > rating - 1 && index < rating) {
       icon = Icon(
         Icons.star_half,
         color: color ?? Theme.of(context).primaryColor,
@@ -32,13 +34,16 @@ class StarRating extends StatelessWidget {
       );
     }
     return InkResponse(
-      onTap: onRatingChanged == null ? null : () => onRatingChanged(index + 1.0),
+      onTap:
+          onRatingChanged == null ? null : () => onRatingChanged(index + 1.0),
       child: icon,
     );
   }
 
   @override
   Widget build(BuildContext context) {
-    return Row(children: List.generate(starCount, (index) => buildStar(context, index)));
+    return Row(
+        children:
+            List.generate(starCount, (index) => buildStar(context, index)));
   }
 }
