+++ lib/templates/middleAssignment/theme.dart
